# Use arrow row format in SortPreservingMerge ~50-70% faster